requirements.txt
streamlit>=1.30.0
crewai>=0.86.0
crewai-tools>=1.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
wikipedia-api>=0.6.0
